Created a new TestCase class for unit tests This class contains `setUp()` and `tearDown()`, which are common to all our tests. Tests will then extend the, and extend the methods if needed.
