+++ tests/Unit/TestCase.php
@@ -1,0 +1,17 @@
+<?php
+namespace WP_Rocket\Tests\Unit;
+
+use PHPUnit\Framework\TestCase as PHPUnitTestCase;
+use Brain\Monkey;
+
+class TestCase extends PHPUnitTestCase {
+	protected function setUp() {
+		parent::setUp();
+		Monkey\setUp();
+	}
+
+	protected function tearDown() {
+		Monkey\tearDown();
+		parent::tearDown();
+	}
+}
